fix model evaluation script and perplexity calculation

+++ lmp/infer/_top_1.py
@@ -70,7 +70,7 @@ class Top1Infer(BaseInfer):
     # Convert to tensor with `dtype == torch.int`.
     # Tensor shape: `(1, S')`.
     # Tensor dtype: `torch.int`.
-    batch_cur_tkids = torch.IntTensor(batch_cur_tkids)
+    batch_cur_tkids = torch.LongTensor(batch_cur_tkids)
 
     # Remove `[eos]` token id since model is not trained to predict tokens
     # after seeing `[eos]`.
